requirements.txt
django
django-filter

# For tests
pyyaml
